Add user configuration playbook
This is not currently tested.

This playbook will create and configure a privileged user. It should
probably be run before the main playbook as the user may need to exist
for some of the configuration. For example, the user should be added to
the informant group when informant is installed.