Add ansible_user to informant group
Will only work if ansible_user has been defined. Should decide on where
the user(s) here should be a variable, or if this playbook should be run
as the user to add to the group.

Ideally the playbook should be user agnostic (as long as that user is
privileged). This probably means a variable is a better idea.